Add test to check output for variables and fix output
