+++ integration-tests/specs/variables/variable_in_path.yaml
@@ -7,3 +7,13 @@ command:
 expected:
   method: GET
   path: /companies/1234
+
+  output: |
+    GET {test-server}/companies/1234
+
+    200 OK 1.1
+    #U# Content-Length: 13
+    #U# Content-Type: text/plain; charset=utf-8
+    #I# Date: Fri, 30 Nov 2018 22:28:48 GMT
+    << Hello world!
+    <<
